Move the main page to a new ListPage component
Fix indentation
