+++ src/App.css
@@ -52,13 +52,6 @@ body {
   flex-grow: 1;
 }
 
-.badge-accent {
-  color: #105c03;
-  background: #19b875;
-  padding: 10px;
-  border-radius: 30;
-}
-
 .footer {
   margin-top: 80px;
   background-color: #23272a;
@@ -78,47 +71,4 @@ body {
   color: white;
   text-align: center;
   background-color: #23272a;
-}
-
-.Locations-header {
-  display: grid;
-  grid-gap: 1rem;
-  padding: 3rem 0;
-}
-
-@media screen and (min-width: 900px) {
-  .Locations-header {
-    grid-template-columns: 1fr 300px;
-    align-items: center;
-  }
-}
-
-.Locations-search {
-  display: block;
-  width: 100%;
-  padding: .8rem 1rem;
-  padding-left: 3rem;
-  border-radius: 1rem;
-  background: #23272a;
-  
-  outline: none;
-  border: none;
-  box-shadow: 0 0 0 2px rgba(255, 2555, 255, .0);
-  transition: .1s ease-in-out box-shadow;
-
-  font-family: inherit;
-  font-size: 1rem;
-
-  /* Heroicons v2.0.12 by Refactoring UI Inc., used under MIT license */
-  background-image: url("data: image/svg+xml, %3Csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 20 20' fill='rgba(255, 2555, 255, .6)' class='w-5 h-5' %3E%3Cpath fill-rule='evenodd' d='M9 3.5a5.5 5.5 0 100 11 5.5 5.5 0 000-11zM2 9a7 7 0 1112.452 4.391l3.328 3.329a.75.75 0 11-1.06 1.06l-3.329-3.328A7 7 0 012 9z' clip-rule='evenodd' /%3E%3C/svg%3E");
-  background-size: 20px;
-  background-repeat: no-repeat;
-  background-position: 1rem center;
-
-  color: white;
-  font-weight: 500;
-}
-
-.Locations-search:focus {
-  box-shadow: 0 0 0 2px rgba(255, 2555, 255, .4);
 }
